main.py/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdfplumber==0.11.0
python-docx==1.1.2
pydantic==2.8.2
spacy==3.7.5
python-multipart==0.0.9
